third_party/blink/web_tests/platform/mac/external/wpt/css/cssom-view/elementsFromPoint-inline-nested-expected.txt

This is a testharness.js-based test.
[FAIL] elementsFromPoint should return all elements under a point, including nested inline elements
  assert_equals: expected 6 but got 3
Harness: the test ran to completion.
